+++ tests/power_of.rs
@@ -10,18 +10,20 @@ const EPSILON: FLOAT = 0.000_001;
 fn test_power_of() -> Result<(), Box<EvalAltResult>> {
     let engine = Engine::new();
 
-    assert_eq!(engine.eval::<INT>("2 ~ 3")?, 8);
-    assert_eq!(engine.eval::<INT>("(-2 ~ 3)")?, -8);
+    assert_eq!(engine.eval::<INT>("2 ** 3")?, 8);
+    assert_eq!(engine.eval::<INT>("(-2 ** 3)")?, -8);
+    assert_eq!(engine.eval::<INT>("2 ** 3 ** 2")?, 512);
 
     #[cfg(not(feature = "no_float"))]
     {
         assert!(
-            (engine.eval::<FLOAT>("2.2 ~ 3.3")? - 13.489_468_760_533_386 as FLOAT).abs() <= EPSILON
+            (engine.eval::<FLOAT>("2.2 ** 3.3")? - 13.489_468_760_533_386 as FLOAT).abs()
+                <= EPSILON
         );
-        assert!((engine.eval::<FLOAT>("2.0~-2.0")? - 0.25 as FLOAT).abs() < EPSILON);
-        assert!((engine.eval::<FLOAT>("(-2.0~-2.0)")? - 0.25 as FLOAT).abs() < EPSILON);
-        assert!((engine.eval::<FLOAT>("(-2.0~-2)")? - 0.25 as FLOAT).abs() < EPSILON);
-        assert_eq!(engine.eval::<INT>("4~3")?, 64);
+        assert!((engine.eval::<FLOAT>("2.0**-2.0")? - 0.25 as FLOAT).abs() < EPSILON);
+        assert!((engine.eval::<FLOAT>("(-2.0**-2.0)")? - 0.25 as FLOAT).abs() < EPSILON);
+        assert!((engine.eval::<FLOAT>("(-2.0**-2)")? - 0.25 as FLOAT).abs() < EPSILON);
+        assert_eq!(engine.eval::<INT>("4**3")?, 64);
     }
 
     Ok(())
@@ -31,26 +33,26 @@ fn test_power_of() -> Result<(), Box<EvalAltResult>> {
 fn test_power_of_equals() -> Result<(), Box<EvalAltResult>> {
     let engine = Engine::new();
 
-    assert_eq!(engine.eval::<INT>("let x = 2; x ~= 3; x")?, 8);
-    assert_eq!(engine.eval::<INT>("let x = -2; x ~= 3; x")?, -8);
+    assert_eq!(engine.eval::<INT>("let x = 2; x **= 3; x")?, 8);
+    assert_eq!(engine.eval::<INT>("let x = -2; x **= 3; x")?, -8);
 
     #[cfg(not(feature = "no_float"))]
     {
         assert!(
-            (engine.eval::<FLOAT>("let x = 2.2; x ~= 3.3; x")? - 13.489_468_760_533_386 as FLOAT)
+            (engine.eval::<FLOAT>("let x = 2.2; x **= 3.3; x")? - 13.489_468_760_533_386 as FLOAT)
                 .abs()
                 <= EPSILON
         );
         assert!(
-            (engine.eval::<FLOAT>("let x = 2.0; x ~= -2.0; x")? - 0.25 as FLOAT).abs() < EPSILON
+            (engine.eval::<FLOAT>("let x = 2.0; x **= -2.0; x")? - 0.25 as FLOAT).abs() < EPSILON
         );
         assert!(
-            (engine.eval::<FLOAT>("let x = -2.0; x ~= -2.0; x")? - 0.25 as FLOAT).abs() < EPSILON
+            (engine.eval::<FLOAT>("let x = -2.0; x **= -2.0; x")? - 0.25 as FLOAT).abs() < EPSILON
         );
         assert!(
-            (engine.eval::<FLOAT>("let x = -2.0; x ~= -2; x")? - 0.25 as FLOAT).abs() < EPSILON
+            (engine.eval::<FLOAT>("let x = -2.0; x **= -2; x")? - 0.25 as FLOAT).abs() < EPSILON
         );
-        assert_eq!(engine.eval::<INT>("let x =4; x ~= 3; x")?, 64);
+        assert_eq!(engine.eval::<INT>("let x =4; x **= 3; x")?, 64);
     }
 
     Ok(())