+++ RELEASES.md
@@ -13,7 +13,12 @@ Bug fixes
 Breaking changes
 ----------------
 
+* In order to be consistent with other scripting languages:
+  * the power/exponentiation operator is changed from `~` to `**`; `~` is now a reserved symbol
+  * the power/exponentiation operator now binds to the right
+  * trigonometry functions now take radians and return radians instead of degrees
 * `Dynamic::into_shared` is no longer available under `no_closure`. It used to panic.
+* `Token::is_operator` is renamed to `Token::is_symbol`.
 
 Enhancements
 ------------
@@ -55,6 +60,7 @@ Enhancements
 * `ahash` is used to hash function call parameters. This should yield speed improvements.
 * `Dynamic` and `ImmutableString` now implement `serde::Serialize` and `serde::Deserialize`.
 * `NativeCallContext` has a new field containing the name of the function called, useful when the same Rust function is registered under multiple names in Rhai.
+* New functions `PI()` and `E()` to return mathematical constants, and `to_radians` and `to_degrees` to convert between radians and degrees.
 
 
 Version 0.19.10